refactor(core/prodtest): move syslog prodtest command to prodtest project
[no changelog]
